Fixed a bunch of warnings

+++ cheat-base/src/cheat-base/thread-safe.h
@@ -19,7 +19,7 @@ public:
 
     virtual ~SafeQueue() { }
 
-    unsigned long size() const {
+    unsigned long long size() const {
         std::lock_guard<std::mutex> lock(mutex_);
         return queue_.size();
     }
